backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
greenlet==3.1.1
asyncpg==0.30.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
redis==5.0.7
celery[redis]==5.4.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
httpx==0.27.0
jinja2==3.1.4
pillow==10.4.0
structlog==24.2.0
sentry-sdk[fastapi]==2.9.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
python-dotenv==1.0.1
